+++ src/execute.ts
@@ -1,7 +1,7 @@
 import ts from 'typescript';
 import { resolve } from 'node:path';
 import { FileService } from './FileService.js';
-import { removeUnusedExport, removeUnusedFile } from './removeExport.js';
+import { removeUnusedExport, removeUnusedFile } from './remove.js';
 import {
   applyCodeFix,
   fixIdDelete,
